+++ src/app/(journal)/index.tsx
@@ -1,15 +1,14 @@
-import { PlatformColor, Pressable, View, useColorScheme } from "react-native";
+import { PlatformColor, Pressable, useColorScheme } from "react-native";
 
-import { Host } from "@expo/ui/swift-ui";
 import { Stack, useRouter } from "expo-router";
 import { SymbolView } from "expo-symbols";
 
-import TemplateList from "@/components/journal/template-list";
+import { TemplateView } from "@/components/template/template-view";
 
 /**
  * ジャーナル画面
  */
-export default function JournalScreen() {
+export default function TemplateScreen() {
   const router = useRouter();
   const colorScheme = useColorScheme();
   const titleColor = colorScheme === "dark" ? "#ffffff" : "#000000";
@@ -26,23 +25,15 @@ export default function JournalScreen() {
           headerLargeTitleEnabled: true,
           headerBackButtonDisplayMode: "minimal",
           headerRight: () => (
-            <Pressable onPress={() => router.push("/(journal)/template")}>
+            <Pressable
+              onPress={() => router.push("/(journal)/template-create")}
+            >
               <SymbolView name="plus" tintColor={PlatformColor("label")} />
             </Pressable>
           ),
         }}
       />
-      <View style={{ flex: 1 }}>
-        <Host
-          style={{
-            flex: 1,
-            backgroundColor: PlatformColor("systemBackground"),
-          }}
-          useViewportSizeMeasurement
-        >
-          <TemplateList />
-        </Host>
-      </View>
+      <TemplateView />
     </>
   );
 }
