+++ src/app/(journal)/template-create.tsx
@@ -1,0 +1,8 @@
+import { TemplateCreateView } from "@/components/template/template-create-view";
+
+/**
+ * テンプレート作成画面
+ */
+export default function TemplateCreateScreen() {
+  return <TemplateCreateView />;
+}
