+++ src/components/template/template-create-view.tsx
@@ -1,0 +1,24 @@
+import { PlatformColor, View } from "react-native";
+
+import { Host, Text } from "@expo/ui/swift-ui";
+import { font, padding } from "@expo/ui/swift-ui/modifiers";
+
+export const TemplateCreateView = () => {
+  return (
+    <View style={{ flex: 1 }}>
+      <Host
+        style={{ flex: 1, backgroundColor: PlatformColor("systemBackground") }}
+        useViewportSizeMeasurement
+      >
+        <Text
+          modifiers={[
+            font({ size: 28, weight: "bold" }),
+            padding({ bottom: 8 }),
+          ]}
+        >
+          テンプレート作成
+        </Text>
+      </Host>
+    </View>
+  );
+};
